Initial import
parse is a collection of parsers for a variety of materials and
chemical modeling and simulation data.

There are a great number of tools that assist with interconversion of
data between chemical formats, e.g. OpenBabel, and for handling
materials and chemical modeling, e.g. the atomistic simulation
envrionment (ASE).